chore: Move maintenance banner to landing page layout
